refactor: 패키지 변경
- news/v1 -> v1/news
